requirements.txt
absl-py==0.10.0
apex @ file:///home/ssr/apex
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828496740/work
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1596809862272/work
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
faiss==1.6.3
faiss-cpu==1.6.3
filelock==3.0.12
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446433964/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206602906/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.10.2
jieba==0.42.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lab==6.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nb-conda==2.2.1
nb-conda-kernels==2.2.3
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook @ file:///tmp/build/80754af9/notebook_1596838602091/work
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
paramiko==2.7.2
path==15.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1594307298074/work
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.6
QtPy==1.9.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
sentence-transformers==0.3.4
sentencepiece==0.1.91
simplejson==3.17.2
six==1.15.0
smart-open==3.0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tfrecord==1.11
threadpoolctl==2.1.0
tokenizers==0.8.1rc1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==3.0.2
txt2tags==3.7
urllib3==1.25.10
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
